pyproject.toml
[project]
name = "rag-hotel"
version = "0.1.0"
description = "this is agent cthat make an action"
authors = [
    {name = "wildr.10"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "openai (>=2.24.0,<3.0.0)",
    "pypdf (>=6.7.4,<7.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "qdrant-client (>=1.17.0,<2.0.0)",
    "langchain-qdrant (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include= "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
